test(widget): migrate Select, Slider, List tests to integration tests
